Fix type of constant nSizeFailure

+++ src/returnval.hpp
@@ -6,7 +6,7 @@ namespace az
 	const int nFailure = 0;
 	const int nTrue = 1;
 	const int nFalse = 0;
-	const int nSizeFailure = -1;
+	const long long int nSizeFailure = -1;
 	const int nCloseSuccess = 0;
 	const int nCloseFailure = -1;
 	const int nReadFailure = -1;
